requirements.txt

# parser for yaml
pyyaml
# log
logbook